requirements.txt
python-dotenv==1.2.2
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
pyodbc==5.3.0